pyproject.toml
[tool.poetry]
name = "insights"
version = "0.1.0"
description = ""
authors = ["Amer Alnajdawi <amer.alnajdawi@treetop-medical.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.14"
streamlit = "^1.35.0"
streamlit-chat = "^0.1.1"
pdfplumber="0.9.0"
openai="0.28"
pydantic = "^2.6.1"
pytesseract = "^0.3.13"
pillow = "^10.4.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
